+++ tests/encoder_tests.py
@@ -10,5 +10,5 @@ def test_encode_to_binary():
 
 def test_decode_to_float():
     asserted_float = encoder.d('1111111111111111111111111111111111111111111111100101100001001010')
-    assert isinstance(asserted_float, long) == True
+    assert isinstance(asserted_float, float) == True
